pull out all camera handling from webxr manager
